refactor: split `Remote...` from local schemas
